Memory Manager code coverage 90%

+++ tests/unit_sample/Qnn/qnn_test.cpp
@@ -113,7 +113,7 @@ typedef struct
     bool bAsync = false;
 } QnnTest_Parameters_t;
 
-static uint32_t s_nodeId = 0;
+static uint8_t s_nodeId = 0;
 
 class QnnTestRunner
 {
